fix: requests will no longer expire outside of normal work hours
This resolves scenarios where a request would come in after hours.
Previously, the request would be marked as expired in the middle of
the night and teammates wouldn't have opportunity to see it unless
they hopped on before 8 AM. Now, if a request comes in after hours
we'll set the request to expire at 9 AM at the start of the next
workday.

closes #78